refactor(discord): extract `ServerMember` into a dedicated module
Simplified imports and improved modularity by moving the `ServerMember` struct and its builder to a new `server_member` module within the Discord connector. This enhances code organization and maintainability.
