+++ nicknamer/src/main.rs
@@ -3,9 +3,9 @@ mod nicknamer;
 use self::nicknamer::commands::names::EmbeddedNamesRepository;
 use self::nicknamer::connectors::discord;
 use self::nicknamer::connectors::discord::serenity::{Context, SerenityDiscordConnector};
+use self::nicknamer::connectors::discord::server_member::ServerMember;
 use crate::nicknamer::commands::nick::{NickService, NickServiceImpl};
 use crate::nicknamer::commands::reveal::{Revealer, RevealerImpl};
-use crate::nicknamer::connectors::discord::ServerMember;
 use log::{LevelFilter, info};
 use log4rs::Config;
 use log4rs::append::console::ConsoleAppender;
